Adding a branchname prepare-commit-msg hook
Git supports several hooks, some of which are client-side. The
prepare-commit-msg hook is run right after a `git commit` call, before
the editor is opened. This allows the initial message to be altered.

This hook will check if the current branch name begins with `bug/`, in
which case it will prepend `[$branchname]` to the commit message. This
makes it easier to create proper commit messages.

http://wiki.phpbb.com/Git#Commit_Messages

For more information refer to the hook source.