[ticket/9778] Replaced "Find a member" with "Enter username"
When clicking on the "Users and groups" tab in the ACP, the user is
able to either enter the username of the member he wants to manage
or click on "Find a member" below the input field. Unfortunately,
the label for the input field is also "Find a member", which leads
some users to the idea that they can search for users using that
input field. By changing the label to "Enter username" this should
be clear to the user.

PHPBB3-9778