[ticket/10013] Changed cache test to use tests/tmp/cache as cache directory
This is the first step of making the test suite use a single directory
for writing files to, this directory being tests/tmp.

The cache test, instead of tests/cache/tmp, now writes to tests/tmp/cache.

Also remove cache directory in setUp method, in case an earlier test run
did not complete successfully and the cache directory was not cleaned up.

Finally, this change makes the cache test take responsibility over the entire
contents of its cache directory.

PHPBB3-10013