[ticket/10101] Add support for native phpass hashes
phpass (the hashing library we use) adds a hash identifier to every
hash. By default this identifier is '$P$'. For some reason we have
changed it to '$H$'.

This patch allows both of them to be used for authentication, so that a
third party system could create users with '$P$' hashes.

PHPBB3-10101