[ticket/9976] Fix link to post when two posts have the same post_time.
When two posts have the same post_time, a link to the first post being on page 1, leads you to the second post with the same timestamp on page two. So in case the post_times are equal, we need to use the post_id to correctly order the posts.

PHPBB3-9976