[ticket/9106] Only display max post size when it is exceeded by an upload.
Do not overload users with information. If a user uploaded an attachment
exceeding the max upload size, but which was under the max post size,
just tell them what the maximum allowed upload size is.

Reserve display of maximum allowed post size for those times when user
uploaded a really huge attachment, this is the same case when they get
instructions on which PHP variables to adjust to make attachments of that
size acceptable.

PHPBB3-9106