[ticket/10144] Check isset() before echoing variables.
This code path is used by loop variables, e.g. loop.S_LAST_ROW,
which are not set for anything but the last row and thus produce
warnings.

This change might be applicable to any variable access in general
since none of the variables referenced in a template need to be set
before the template is evaluated.

PHPBB3-10144