[ticket/8173] Removed some redundant code in the BBCode JavaScript.
The following code can be found in the ACP, when changing a users
signature:

t: '{ BBCODE_T_HELP }',

The code that uses it and the code that produces it were changed or
removed a while ago, but this code remained. In
2c4c1f9, the accesskey for the list
item button was changed to y from t, explaining the origin of the
redundant code - although it wasn't being used then, either.

This commit also changes the helpline code of "[*]" to y, as the access
key is y, so this makes it more consistant.

PHPBB3-8173