[ticket/10414] Skip functional tests on PHP 5.2 - requires PHPUnit 3.6 on 5.2
Tests still execute correctly using PHPUnit 3.5 on PHP 5.3 and above. The php
version limitation for a directory was added in PHPUnit 3.6. A separate test
suite is required because the functional tests are in the whitelisted tests
directory. The base test for functional testing is only included in bootstrap
in versions 5.3 and above.

PHPBB3-10414