[ticket/10439] Fixed an inefficiency when fetching rows from memory
When fetching SQL rows from memory, a sizeof would be called for each
row fetched, which when working with large sets of data (I encountered
it with ban lists) takes its toll. Now sizeof is called once when the
data is retrieved and stored in a variable that can be used when
determining the number of rows in the query set.

PHPBB3-10439