[ticket/10407] Fix check for empty image paths in convertor
This applies to avatar_path,avatar_gallery_path,smilies_path and
upload_path.
Currently, the convertor gets each path from the config table and adds a
trailing slash. If there is no entry in the config table for the path, the
path can never be empty because of the added trailing slash. This patch
will check the path without the trailing slash to see if the path is
actually empty.

PHPBB3-10407