[ticket/8652] Sending 2 emails on 2 replies
This is a fix for the problem of the system sending 2 emails when there are 2
replies to a topic where the user is subscribed in a topic and the forum that
contains the topic.
This simple fix seems to solve the problem. In simple tests I made it shows
it does.

PHPBB3-8652