[ticket/8652] Sending 2 emails on 2 replies
Changed the comment by request to a more verbose version.
This version is supposed to explain better which problem is it trying to solve
and why that code is not a bad idea to solve the problem (in therms of
performace hit caused by the extra DB query which is actually only a performance
hit if no user is subscribed to the forum that includes that thread,
which is something very rare).

PHPBB3-8652