[ticket/10586] Change the interface to an abstract class
This allows the common phpBB objects to be automatically accessible to
extensions without the author having to globalize and assign each one
himself. This is better because it also gives purpose to the
phpbb_extension_controller class; instead of just being the way to
ensure a handle() method is present, it also does this work for us.

PHPBB3-10586