[ticket/10724] Preparing the previous code to receive the new one
The previous code had a problem. At least, with php 5.3 and php 5.4 in windows,
php lint was returning the status code of 0 when there were syntax errors
instead of -1. This was bypassing the check that is made for syntax errors.

This prepares this hook for the upcoming change to add the check for the
whitespace characters check by changing variable names and by fixing the
mensioned problem (bug?)

PHPBB3-10724