[ticket/10724] Adding the check for trailling spaces in lines
This adds the check for trailling whitespace characters in lines.
It checks for all the lines that contain, at least, one whitespace character
and then it prints them. While pritting them, it shows the same information as
the php lint. The filename and the line where the finding occured. Additionally,
because grep was used, it shows the text of the line where the situation
happened

PHPBB3-10724