[ticket/10760] Use externally specified PHP_BIN if any.
We have a PHP_BIN variable but we always set it unconditionally,
therefore in order to use a different php binary one had to
modify the hook script.

Instead set PHP_BIN if it is not set.

Now one can set PHP_BIN in their environment to use a non-default
php binary.

PHPBB3-10760