[ticket/10780] Move colons from template files to language files.
For now, colons are hardcoded in template files (adm, prosilver & subsilver2). In some languages, a space is needed before it or it's possible that another symbol is used. For this reason, colons are better located in the language files. *_COLON language key has been added when a language key is also used in another place without a colon.

PHPBB3-10780