[feature/only-op-can-view] Remove the possibility of an SQL "OR 1=1" condition
Rewored a (little) problematic SQL query in order to optimize it.
Now the condition that was always being calculated, only calculates when the
current user does not have f_read_other rights for that specific forum.
This should improve this code's performance by a little.
Also, moved the $user->data['user_id'] check to first place. As it is the most
probable thing that will say the final true to the condition (minor improvement)

PHPBB3-10623