[feature/revisions] New revision diff handling class, more work on diff
I was trying to find a good way to display the diff when I realized it
might work to do something similar to the GitHub Compare View. It's getting
late, so I'll leave off here for now, but tomorrow I plan on restructuring
it to look more like that.

PHPBB3-10755