[feature/revisions] Created API class for diff engines to plug into
This allows for us to use different diff engines as needed. No, I don't
expect that it would really ever need to happen for anyone to change the
diff engine, but this way they could create a class to interface between
the diff engine and phpBB without having to touch the logic itself. This
also makes it easier later one when I actually have to find or make an
appropriate diff engine since FineDiff may not be a mature enough library.

This commit also includes some work on a different form than confirm box.
confirm_box() is a good generic confirmation method but it doesn't allow
as much flexibility as a custom form. This way I can display the changes
to be made on the confirmation page.

PHPBB3-10755