[feature/revisions] Addressed various PR comments
I split up the cron tasks, and will continue to look at optimizing them.
I also removed the inline assignments in if() statements, as well as a
few other stylistic changes, including splitting up long conditional
statements and commenting them to make them clearer.

In other news, I also began re-implementing revision comparison, but it
is still not nearly ready. As with the rest of it, this is a WIP.

PHPBB3-10755