[feature/attachment-improvements] Incorporated comments from the PR
Reverted changes to user.php since they are no longer needed.

Created a new cron job to clear the plupload temporary directory.

The file token used by plupload is now hashed and salted for increased
security.

Various changes to the install script and database schema have been made to
incorporate the above changes.

Uploading chunked files is still not working as the mimetype is lost over the
chunks and so the file fails preliminary image checks. The real file name as
belonging to the user is also lost during upload.

PHPBB3-10929