[ticket/10962] Request class doesn't return correct array default value
When using the request class the second parameter defines the "default" value,
if you pass a string or an integer as second parameter and the parameter isn't
set in the `$_REQUEST` array the class will neatly return the value you've
passed as default.
However when passing an array the class will return an empty array rather than
the one you've passed as default.
This behavior is unexpected and undocumented in the docblock.

PHPBB3-10962