[feature/revisions] Allow viewing a single revision
Will need to expand on that page with more information about
the revision. At the moment it just displays it as if it were
the current post.

Also fixed a bug in which the post ID was not provided by
the $post_id variable in viewtopic when the ?p= parameter
was not given in the URL. Instead, it needs to use
$row['post_id']

PHPBB3-10755