[feature/auth-refactoring] Handle getting required data
This allows providers, such as OpenID, to request a username or email
address from the user due to an already existing e-mail, or and invalid
or already used username.
No providers yet have support for handling the request.
Fixes OpenID login issues.
Assigns a random password to all created accounts that are made without
a password set by a provider.

PHPBB3-9734