[feature/auth-refactoring] Document and Clean Auth Code
Mostly just a lot of documentation that was previously missing. Much of it
still needs to be fleshed out in more detail; however, the basic
documentation structure is in place.

Renames phpbb_auth_provider_custom_registration_interface to
phpbb_auth_provider_registration_interface as it is more accurate as to
the function of the interface in that only providers with this interface
should draw a registration interface.

Changes ucp_register to only request providers with the
phpbb_auth_provider_registration_interface attempt to draw a registration
template.

PHPBB3-9734