Update phpBB/includes/db/mssqlnative.php
For some reason, all errors are just flipped on before connecting to the database, despite the system as a whole having a different setting for displayable errors. Had to add & ~E_STRICT in PHP 5.4.5 to suppress Strict Standards messages, but I would assume that the db piece shouldn't be involved with setting error_reporting at all.