[ticket/11090] Cron dies with suppressed fatal error
Common.php includes config.php twice, triggering the error handler
to be called (because of suppressed redefined constants). This error
makes the error handler try to log to the database, which the
connection for does not exist yet, leading to a fatal error.

This patch fixes the issue by making sure that the error handler will
only perform logging for errors that occur after common.php has been
included.

PHPBB3-11090