[ticket/8713] Revert changes to ucp_profile, ucp_register and acp_users.
Currently only auth_db is supported there and the password needs to be trimmed
for compatibility because user_password stores

  phpbb_hash(htmlspecialchars(trim($password)))

Setting passwords for other auth modules is currently not supported.

Once setting/changing passwords is supported by auth plugins, the
untrimmed_variable() should be used here and the result should be passed
to the auth plugin.

PHPBB3-8713