[ticket/11103] Trying to fix an odd issue with unread status on approved posts
From a recent change, when your posts/topics are approved, they will be
marked read automatically because you've read the topic/post already.

To change that I've forced the notification to be marked unread and
attempted to reset the read status on the post/topic to be unread before
the post that was approved.

This does not seem to work so well and I don't know of any way this can
really be properly fixed, so the code I was working on I've commented out.

For now, users will just need to manually mark these types of notifications
as read. I cannot think of a way for this to be fixed without running
two additional queries on every viewtopic.

PHPBB3-11103