[ticket/10879] Fixup use of arrow icons in prosilver
The current class names for the arrow icons in prosilver are ambiguous.
This will rename the class name to arrow-(+direction), i.e. arrow-right.
Additionally, it will get rid of the arrow icon next to the attachment
link on the posting editor. If you upload a file with a long filename
the filename will partially cover the arrow_right.gif and the icon isn't
needed for indicating that it's a link to the attached file.

PHPBB3-10879