[feature/autosave] More work on the autosave.
Autosaves should now save per window on pages with posting. When
loading a page, all autosaves of that post will be prompted for,
organised by the time at which they were created.

Added a language string and a few template variables required by
javascript for calculating the time correctly for the user's
timezone.

PHPBB3-11185