[ticket/10716] Collect standard error from executed php process.
php executes everything via a shell. The standard error of this
top level shell is not captured by exec/shell_exec/popen/etc.
and there is no way to capture it. proc_open might work but it
is a nightmare to use and without multiplexing reads from
standard error and standard output it can deadlock.

Thus the solution in this commit. Put the command into a subshell
and redirect standard error to standard output for the subshell.

PHPBB3-10716