[ticket/10805] Clear loading alert timeout after ajax request finished
The timeout for the "request timed out" popup should be cleared if it
finished. Since it is currently not cleared, the timeout alert appears as
an extra overlay if another ajaxified function is ran within 5 seconds of
the initial function call. This patch will take care of clearing the
timeout if either the success (function return_handler()) or error
(function error_handler()) functions are called.

PHPBB3-10805