[ticket/11278] Comment out the code for dropping the Q&A tables
Due to a bug, vanilla phpbb could not create captcha tables in 3.0.8 on
firebird. It was possible for board administrators to adjust the code to
work. If code was manually adjusted by board administrators, index names
would not be the same as what 3.0.9 and newer expect. This code fragment
drops captcha tables, destroying all entered Q&A captcha configuration,
such that when Q&A is configured next the respective tables will be
created with corrent index names.

If you wish to preserve your Q&A captcha configuration, you can manually
rename indexes to the currently expected name:
 phpbb_captcha_questions_lang_iso  => phpbb_captcha_question_lang
 phpbb_captcha_answers_question_id => phpbb_captcha_answers_qid

Again, this needs to be done only if a board was manually modified to fix
broken captcha code.

PHPBB3-11278