[ticket/11295] Drop varchar_ci type when purging extras on postgres.
This fixes the database recreation path when initially database
drop is attempted and it fails, and the code drops tables.
varchar_ci was not dropped in that situation, creating a failure
when schema was subsequently loaded.

Example reproduce case:

1. Create users A and B.
2. Create a test database owned by user A.
3. Tell test suite to use user B and database owned by user A.
4. Run the test suite (twice).

Database drop will fail, tables will be dropped, varchar_ci will
not be dropped, causing a failure when loading schema.

PHPBB3-11295