[ticket/11030] Adding a new style class to file upload in Prosilver.
In Prosilver, browse button of file upload element in post attachment
shows a cursor on mouse over. That was because a CSS property is added
with a common class name "inputbox". I added a new CSS class for file
upload element and included it in file upload form template for Prosilver.

PHPBB3-11030