[ticket/6723] Show info that message has been deleted before delivery
If the message has been deleted before the delivery it will still show in
the message history, although with an empty message. This patch will
change that behavior to showing the info that it has been deleted before
it was delivered in red letters. The red letters ensure that users won't
mistake the info for the actual message.

PHPBB3-6723