[ticket/11340] Using AJAX Delete alters pagination.
When deleting a post from a particular page, posts should shift
back and last post from next page should come to end of the current
page. But it does not happen. When clicking the next page, the
first post that was on the next page before the post was deleted
is now on the current page. Therefore When clicking on next page
user will miss a post. Added a new ajax call to get the post from
next page and add it to the end of the current page by using
post_delete callback.

PHPBB3-11340